pyproject.toml
[project]
authors = [{ name = "Volker L.", email = "ligno.blades@gmail.com" }]
dependencies = [
  "pyarrow>=12.0.1",
  "s3fs>=2023.6.0",
  "duckdb>=0.10.0",
  "tqdm>=4.65.0",
  "joblib>=1.3.2",
  "pandas>=2.0.3",
  "pendulum>=2.1.2", #"msgspec>=0.18.4",
  "munch>=4.0.0",
  "sqlglot>=23.11.2",
  "sqlparse>=0.5.0",
  "psutil>=6.0.0",
  "pyyaml>=6.0.2",
  "loguru>=0.7.2",
  "orjson>=3.10.12",
  "fsspec-utils[full]>=0.1.10",
  "rich>=14.1.0",
]
description = "poor man´s data lake"
#homepage = "https://github.com/legout/pydala2"
license = { text = "MIT" }
name = "pydala2"
readme = "README.md"
#repository = "https://github.com/legout/pydala2"
requires-python = ">= 3.11"
version = "0.20.0"

[project.optional-dependencies]
legacy = ["polars-lts-cpu>=0.20.4"]
standard = ["polars>=0.20.4"]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling==1.26.3", "hatch-vcs"]

# [build-system]
# build-backend = "setuptools.build_meta"
# requires = ["setuptools>=42", "wheel"]

[tool.rye]
dev-dependencies = [
  "ipython>=8.26.0",
  "isort>=5.13.2",
  "deltalake>=0.20.2",
  "datafusion>=42.0.0",
  "ibis-framework[duckdb,polars]>=9.5.0",
  "ruff>=0.7.1",
  "jupyterlab>=4.3.0",
]
managed = true

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["pydala"]

[dependency-groups]
dev = [
    "anyio>=4.9.0",
    "ipython>=9.5.0",
    "jupyterlab>=4.4.7",
    "marimo>=0.15.3",
]
